requirements.txt
amqp==5.1.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
kombu==5.2.4
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==4.21.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
vine==5.0.0
